Support json/xml request bodies
Also fix the handling of content-type/length in requests for the
purposes of sending these headers to the helper and for setting the
respective tags. These are available in $_SERVER['CONTENT_...'], not
$_SERVER['HTTP_CONTENT_...']
