+++ appsec/src/extension/ddtrace.h
@@ -63,7 +63,7 @@ struct _ddtrace_user_req_listeners {
     int priority;
     zend_array *nullable (*nonnull start_user_req)(
         ddtrace_user_req_listeners *nonnull self, zend_object *nonnull span,
-        zend_array *nonnull variables);
+        zend_array *nonnull variables, zval *nullable rbe_zv);
     zend_array *nullable(*nonnull response_committed)(
         ddtrace_user_req_listeners *nonnull self, zend_object *nonnull span,
         int status, zend_array *nonnull headers, zval *nullable entity);
