+++ appsec/src/extension/entity_body.stub.php
